Update button name + style + loader
