+++ src/components/spinner/spinner.module.css
@@ -2,8 +2,6 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 40px;
-  padding: var(--s);
 }
 
 .spinner {
