+++ src/routes/App.css
@@ -65,6 +65,10 @@ button.drawer-btn {
 }
 
 /* =========== BUTTON =========== */
+button[type="submit"] {
+  min-width: 140px!important;
+}
+
 button.btn-icon {
   display: flex;
   align-items: center;
@@ -75,6 +79,11 @@ button.action, button.btn-icon {
   min-height: 40px;
   color: var(--primary-dark);
   background-color: var(--action);
+}
+
+button.action:disabled, button.btn-icon:disabled {
+  cursor: unset;
+  pointer-events: none;
 }
 
 button.action:hover, button.btn-icon:hover {
